Make grammar analyser prettier
